Login still lacking token and was not able to do donate to platform

+++ KMS/KMS-war/src/java/ejb/session/stateless/UserSessionBean.java
@@ -1,15 +1,17 @@
 package ejb.session.stateless;
 
+import Exception.AffiliatedUserExistException;
 import Exception.DuplicateEmailException;
+import Exception.DuplicateFollowRequestException;
 import Exception.DuplicateTagInProfileException;
 import Exception.InvalidLoginCredentialException;
 import Exception.InvalidUUIDException;
 import Exception.NoResultException;
 import Exception.UserNotFoundException;
 import entity.FollowRequestEntity;
-import entity.MaterialResourceAvailableEntity;
 import entity.TagEntity;
 import entity.UserEntity;
+import java.util.ArrayList;
 import java.security.SecureRandom;
 import java.util.List;
 import java.util.Properties;
@@ -27,6 +29,7 @@ import javax.persistence.NonUniqueResultException;
 import javax.persistence.PersistenceContext;
 import javax.persistence.Query;
 import util.enumeration.AccountPrivacySettingEnum;
+import util.enumeration.UserTypeEnum;
 import util.security.CryptographicHelper;
 
 /**
@@ -211,6 +214,16 @@ public class UserSessionBean implements UserSessionBeanLocal {
     }
 
     @Override
+    public List<TagEntity> getSDGsForProfile(long userId) throws UserNotFoundException {
+        UserEntity user = em.find(UserEntity.class, userId);
+
+        if (user == null) {
+            throw new UserNotFoundException("User not found.");
+        }
+        return user.getSdgs();
+    }
+
+    @Override
     public void addSDGToProfile(long userId, long tagId) throws NoResultException, DuplicateTagInProfileException {
         UserEntity user = em.find(UserEntity.class, userId);
         TagEntity tag = em.find(TagEntity.class, tagId);
@@ -304,9 +317,7 @@ public class UserSessionBean implements UserSessionBeanLocal {
     }
 
     @Override
-    public void followUser(Long toUserId, Long fromUserId) throws UserNotFoundException {
-        System.out.println(toUserId);
-        System.out.println(fromUserId);
+    public FollowRequestEntity followUser(Long toUserId, Long fromUserId) throws UserNotFoundException, DuplicateFollowRequestException {
         UserEntity toUser = em.find(UserEntity.class, toUserId);
         UserEntity fromUser = em.find(UserEntity.class, fromUserId);
         if (toUser == null || fromUser == null) {
@@ -315,31 +326,74 @@ public class UserSessionBean implements UserSessionBeanLocal {
         if (toUser.getAccountPrivacySetting().equals(AccountPrivacySettingEnum.PUBLIC)) {
             fromUser.getFollowing().add(toUser);
             toUser.getFollowers().add(fromUser);
+            return null;
         } else {
-            FollowRequestEntity followRequestEntity = new FollowRequestEntity();
-            em.persist(followRequestEntity);
-            followRequestEntity.setFrom(fromUser);
-            followRequestEntity.setTo(toUser);
+            Query q = em.createQuery("SELECT f FROM FollowRequestEntity AS f WHERE f.from.userId = :from AND f.to.userId = :to");
+            q.setParameter("from", fromUserId);
+            q.setParameter("to", toUserId);
+            try {
+                FollowRequestEntity f = (FollowRequestEntity) q.getSingleResult();
+                throw new DuplicateFollowRequestException("Follow request already sent!");
+            } catch (javax.persistence.NoResultException e) {
+                FollowRequestEntity followRequestEntity = new FollowRequestEntity();
+                followRequestEntity.setFrom(fromUser);
+                followRequestEntity.setTo(toUser);
+
+                fromUser.getFollowRequestMade().add(followRequestEntity);
+                toUser.getFollowRequestReceived().add(followRequestEntity);
+                em.persist(followRequestEntity);
+                em.flush();
+                return followRequestEntity;
+            }
         }
     }
 
     @Override
     public void acceptFollowRequest(Long toUserId, Long fromUserId) throws NoResultException, UserNotFoundException {
-        Query q = em.createQuery("SELECT f FROM FollowRequestEntity AS f WHERE f.from=:from AND f.to=:to");
+        Query q = em.createQuery("SELECT f FROM FollowRequestEntity AS f WHERE f.from.userId = :from AND f.to.userId = :to");
+        q.setParameter("from", fromUserId);
+        q.setParameter("to", toUserId);
+        FollowRequestEntity f = (FollowRequestEntity) q.getSingleResult();
+        if (f == null) {
+            throw new NoResultException("Follow request not found");
+        }
+        UserEntity toUser = em.find(UserEntity.class, toUserId);
+        UserEntity fromUser = em.find(UserEntity.class, fromUserId);
+        if (toUser == null || fromUser == null) {
+            throw new UserNotFoundException("User not found");
+        }
+        fromUser.getFollowRequestMade().remove(f);
+        toUser.getFollowRequestReceived().remove(f);
+        f.setTo(null);
+        f.setFrom(null);
+        em.remove(f);
+        if (!fromUser.getFollowing().contains(toUser)) {
+            fromUser.getFollowing().add(toUser);
+        }
+        if (!toUser.getFollowers().contains(fromUser)) {
+            toUser.getFollowers().add(fromUser);
+        }
+    }
+
+    @Override
+    public void rejectFollowRequest(Long toUserId, Long fromUserId) throws NoResultException, UserNotFoundException {
+        Query q = em.createQuery("SELECT f FROM FollowRequestEntity AS f WHERE f.from.userId = :from AND f.to.userId = :to");
         q.setParameter("from", fromUserId);
         q.setParameter("to", toUserId);
         FollowRequestEntity f = (FollowRequestEntity) q.getSingleResult();
         if (f == null) {
             throw new NoResultException("follow request not found");
         }
-        em.remove(f);
         UserEntity toUser = em.find(UserEntity.class, toUserId);
-        UserEntity fromUser = em.find(UserEntity.class, toUserId);
+        UserEntity fromUser = em.find(UserEntity.class, fromUserId);
         if (toUser == null || fromUser == null) {
             throw new UserNotFoundException("User not found");
         }
-        fromUser.getFollowing().add(toUser);
-        toUser.getFollowers().add(fromUser);
+        fromUser.getFollowRequestMade().remove(f);
+        toUser.getFollowRequestReceived().remove(f);
+        f.setTo(null);
+        f.setFrom(null);
+        em.remove(f);
     }
 
     @Override
@@ -380,13 +434,78 @@ public class UserSessionBean implements UserSessionBeanLocal {
 
 
     @Override
-    public List<UserEntity> getAllUsers() throws NoResultException{
+    public List<UserEntity> getAllUsers() throws NoResultException {
         Query q = em.createQuery("SELECT u FROM UserEntity u");
         List<UserEntity> users = q.getResultList();
+        for (UserEntity userEntity : users) {
+            userEntity.getFollowers().size();
+            userEntity.getFollowing().size();
+        }
         return users;
     }
+
+    @Override
+    public List<UserEntity> getAffiliatedUsers(Long userId) throws UserNotFoundException {
+        UserEntity user = em.find(UserEntity.class, userId);
+        if (user == null) {
+            throw new UserNotFoundException("User not found");
+        }
+        List<UserEntity> users = new ArrayList<>();
+        if (user.getUserType() == UserTypeEnum.INDIVIDUAL) {
+            users = user.getAffiliatedInstitutes();
+        } else if (user.getUserType() == UserTypeEnum.INSTITUTE) {
+            users = user.getAffiliatedIndividuals();
+        }
+        for (UserEntity userEntity : users) {
+            userEntity.getFollowers().size();
+            userEntity.getFollowing().size();
+        }
+        return users;
+    }
+
+    @Override
+    public void addAffiliatedUser(Long userId, Long affiliatedToAddUserId) throws AffiliatedUserExistException, UserNotFoundException {
+        UserEntity user = em.find(UserEntity.class, userId);
+        UserEntity affiliatedUserToAdd = em.find(UserEntity.class, affiliatedToAddUserId);
+        if (user == null || affiliatedUserToAdd == null) {
+            throw new UserNotFoundException("User not found");
+        }
+        List<UserEntity> affiliatedUsers = new ArrayList<>();
+        if (user.getUserType() == UserTypeEnum.INDIVIDUAL) {
+            affiliatedUsers = user.getAffiliatedInstitutes();
+        } else if (user.getUserType() == UserTypeEnum.INSTITUTE) {
+            affiliatedUsers = user.getAffiliatedIndividuals();
+        }
+
+        if (!affiliatedUsers.contains(affiliatedUserToAdd)) {
+            affiliatedUsers.add(affiliatedUserToAdd);
+        } else{
+            throw new AffiliatedUserExistException("User is already affiliated.");
+        }
+    }
     
-    public UserEntity updateUser(UserEntity updatedUser) throws UserNotFoundException, DuplicateEmailException {
+    @Override
+    public void removeAffiliatedUser(Long userId, Long affiliatedToRemoveUserId) throws NoResultException, UserNotFoundException {
+        UserEntity user = em.find(UserEntity.class, userId);
+        UserEntity affiliatedUserToRemove = em.find(UserEntity.class, affiliatedToRemoveUserId);
+        if (user == null || affiliatedUserToRemove == null) {
+            throw new UserNotFoundException("User not found");
+        }
+        List<UserEntity> affiliatedUsers = new ArrayList<>();
+        if (user.getUserType() == UserTypeEnum.INDIVIDUAL) {
+            affiliatedUsers = user.getAffiliatedInstitutes();
+        } else if (user.getUserType() == UserTypeEnum.INSTITUTE) {
+            affiliatedUsers = user.getAffiliatedIndividuals();
+        }
+
+        if (affiliatedUsers.contains(affiliatedUserToRemove)) {
+            affiliatedUsers.remove(affiliatedUserToRemove);
+        } else{
+            throw new NoResultException("User is not affiliated.");
+        }
+    }
+
+    public UserEntity updateUser(UserEntity updatedUser) throws UserNotFoundException, DuplicateEmailException, NoResultException {
         UserEntity user = em.find(UserEntity.class, updatedUser.getUserId());
         if (user == null) {
             throw new UserNotFoundException("User not found");
@@ -405,6 +524,14 @@ public class UserSessionBean implements UserSessionBeanLocal {
         user.setProfilePicture(updatedUser.getProfilePicture());
         user.setAccountPrivacySetting(updatedUser.getAccountPrivacySetting());
 
+        for (int i = 0; i < updatedUser.getSdgs().size(); i++) {
+            TagEntity tag = em.find(TagEntity.class, updatedUser.getSdgs().get(i).getTagId());
+            if (tag == null) {
+                throw new NoResultException("SDG tag not found.");
+            }
+        }
+        user.setSdgs(updatedUser.getSdgs());
+
         return user;
     }
 
@@ -414,32 +541,37 @@ public class UserSessionBean implements UserSessionBeanLocal {
         if (user == null) {
             throw new UserNotFoundException("User not found");
         }
+        user.getFollowers().size();
         return user.getFollowers();
     }
-    
+
     @Override
     public List<UserEntity> getFollowing(long userId) throws UserNotFoundException {
         UserEntity user = em.find(UserEntity.class, userId);
         if (user == null) {
             throw new UserNotFoundException("User not found");
         }
+        user.getFollowing().size();
         return user.getFollowing();
     }
-    
+
     @Override
-    public List<MaterialResourceAvailableEntity> getMaterialRequestAvailable(long userId) throws UserNotFoundException {
+    public List<FollowRequestEntity> getFollowRequestReceived(Long userId) throws UserNotFoundException {
         UserEntity user = em.find(UserEntity.class, userId);
         if (user == null) {
             throw new UserNotFoundException("User not found");
         }
-        return user.getMras();
+        user.getFollowRequestReceived().size();
+        return user.getFollowRequestReceived();
     }
 
-    
     @Override
-    public List<UserEntity> retrieveAllUser() throws UserNotFoundException{
-        Query query = em.createQuery("SELECT u FROM UserEntity U");
-        return query.getResultList();
+    public List<FollowRequestEntity> getFollowRequestMade(Long userId) throws UserNotFoundException {
+        UserEntity user = em.find(UserEntity.class, userId);
+        if (user == null) {
+            throw new UserNotFoundException("User not found");
+        }
+        user.getFollowRequestMade().size();
+        return user.getFollowRequestMade();
     }
-
 }
